move files to /src folder, etc.
