+++ package.json
@@ -9,6 +9,8 @@
   },
   "dependencies": {
     "@apollo/client": "^3.5.6",
+    "@fontsource/inter": "^4.5.1",
+    "@fontsource/jetbrains-mono": "^4.5.0",
     "@solana/wallet-adapter-base": "^0.9.2",
     "@solana/wallet-adapter-react": "^0.15.2",
     "@solana/wallet-adapter-wallets": "^0.14.2",
